Added appsettings.json file support.
Secrets, IE: things that should NOT be tracked in a git repository, are to be put in a file called appsettings.{module}.secret.json.

The repository will not track any .secret.json files.

This is also only implemented for the connection string at the moment and only supported read operations. More to be expanded upon later.